+++ src/main/java/subway/line/LineController.java
@@ -2,6 +2,7 @@ package subway.line;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.line.section.SectionsUpdateRequest;
 
 import java.net.URI;
 import java.util.List;
