+++ src/main/java/subway/line/section/Section.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.section;
 
 import subway.station.Station;
