+++ src/main/java/subway/line/section/SectionsUpdateRequest.java
@@ -1,4 +1,4 @@
-package subway.line;
+package subway.line.section;
 
 public class SectionsUpdateRequest {
     private Long downStationId;
